Add logging about the environments and current environment
I've been struggling with Cypress tests in CI - this middleware seems to be running when Cypress runs, and this middleware is forwarding it to https even when running in localhost, which is confusing since the NODE_ENV shouldn't be set to production in that case.  Going to check it out.